+++ management/server/integrations/integrated_validator/interface.go
